refactor(JWT) JWt Global 클래스 리펙토링 -> 역할에 따른 클래스 분할
