+++ src/main/java/org/com/dungeontalk/global/security/JwtExtractor.java
@@ -1,0 +1,18 @@
+package org.com.dungeontalk.global.security;
+
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.stereotype.Service;
+
+@Service
+public class JwtExtractor {
+
+    // 유저 요청으로 부터 엑세스 토큰 추출
+    public String extractAccessToken(HttpServletRequest request) {
+        String authorizationHeader = request.getHeader("Authorization");
+
+        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            return authorizationHeader.substring(7);
+        }
+        return null;
+    }
+}
