+++ src/main/java/org/com/dungeontalk/global/security/JwtProvider.java
@@ -1,0 +1,72 @@
+package org.com.dungeontalk.global.security;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.security.Keys;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import java.nio.charset.StandardCharsets;
+import java.security.Key;
+import java.util.Date;
+
+@Slf4j
+@Service
+public class JwtProvider {
+    @Value("${jwt.secret}")
+    private String SECRET_KEY;
+
+    @Value("${jwt.accessexpiration}")
+    private long ACCESS_TOKEN_EXPIRATION_TIME;
+
+    @Value("${jwt.refreshexpiration}")
+    private long REFRESH_TOKEN_EXPIRATION_TIME;
+
+    // 엑세스 토큰 생성 v2
+    public String generateAccessToken(String id, String name, String nickName) {
+        Key key = Keys.hmacShaKeyFor(SECRET_KEY.getBytes(StandardCharsets.UTF_8));
+
+        return Jwts.builder()
+                .setSubject(id) // 보편적으로 PK가 subject
+                .claim("name", name)
+                .claim("nickName", nickName)
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(System.currentTimeMillis() + ACCESS_TOKEN_EXPIRATION_TIME))
+                .signWith(key, SignatureAlgorithm.HS256) // 비대칭 키 적용
+                .compact();
+    }
+
+    // 리프레시 토큰 생성 v2
+    public String generateRefreshToken(String id) {
+        Key key = Keys.hmacShaKeyFor(SECRET_KEY.getBytes(StandardCharsets.UTF_8));
+
+        return Jwts.builder()
+                .setSubject("id") // PK를 subject로
+                .claim("type", "refresh") // 토큰 타입 명시
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(System.currentTimeMillis() + REFRESH_TOKEN_EXPIRATION_TIME))
+                .signWith(key, SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    // 토큰에서 클레임 추출
+    public Claims extractClaims(String token) {
+        return Jwts.parser() // JWT 파서 객체 생성
+                .setSigningKey(SECRET_KEY)
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    public boolean validateToken(String token) {
+
+        try {
+            Claims claims = extractClaims(token);
+            return !claims.getExpiration().before(new Date());
+        } catch (Exception e) {
+            log.error("Token 유효성 검증 실패 : {}", token, e);
+            return false;
+        }
+    }
+}
